+++ src/main/java/com/pvt/demo/repository/EventInstanceRepository.java
@@ -2,4 +2,4 @@ package com.pvt.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.pvt.demo.model.EventInstance;
 
-public interface EventInstanceRepository extends JpaRepository<EventInstance, Integer> {}
+public interface EventInstanceRepository extends JpaRepository<EventInstance, Long> {}
