+++ src/main/java/com/pvt/demo/repository/RecurringEventRepository.java
@@ -3,4 +3,4 @@ package com.pvt.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.pvt.demo.model.RecurringEvent;
 
-public interface RecurringEventRepository extends JpaRepository<RecurringEvent, Integer> {}
+public interface RecurringEventRepository extends JpaRepository<RecurringEvent, Long> {}
