+++ src/main/java/com/pvt/demo/repository/UserRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pvt.demo.model.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {}
+public interface UserRepository extends JpaRepository<User, Long> {}
